- get rid of the warning.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@5750 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/interactive/omi_Transfer.cpp
@@ -296,7 +296,7 @@ void doTransfer(long nStates, long nAlgebraic, long nParameters) {
 		//if(debugTransfer) cout << "SimResDataForw_from_Transfer address: " << p_SimResDataForw_from_Transfer << endl;
 		//printSSDTransfer(nStates, nAlgebraic, nParameters);
 		sendMessageToClientGUI(nStates, nAlgebraic, nParameters);
-		delay(get_stepSize() * 1000); //TODO 20100427 pv The sending frequency should depend on the real time, **soft real time
+		delay((unsigned int)(get_stepSize() * 1000)); //TODO 20100427 pv The sending frequency should depend on the real time, **soft real time
 	}
 }
